server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
supabase==2.9.0
python-jose[cryptography]==3.3.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
google-generativeai==0.8.3
websockets==13.1
httpx==0.27.2

